+++ crates/selene-gql/src/runtime/pipeline/catalog/mod.rs
@@ -442,6 +442,7 @@ fn render_gql_type(ty: &GqlType) -> String {
         GqlType::Float32 => "FLOAT32".to_owned(),
         GqlType::Float64 => "FLOAT64".to_owned(),
         GqlType::Bytes | GqlType::Binary | GqlType::VarBinary => "BYTES".to_owned(),
+        GqlType::Uuid => "UUID".to_owned(),
         GqlType::ZonedDateTime => "ZONED DATETIME".to_owned(),
         GqlType::LocalDateTime => "LOCAL DATETIME".to_owned(),
         GqlType::Date => "DATE".to_owned(),